Move loading spinner to components folder
