+++ src/routes/TaskLists/TaskListPage.jsx
@@ -32,7 +32,7 @@ import { modifyRoRoPassengersTaskList } from '../../utils/roroDataUtil';
 // Components/Pages
 import ClaimButton from '../../components/ClaimTaskButton';
 import ErrorSummary from '../../govuk/ErrorSummary';
-import LoadingSpinner from '../../forms/LoadingSpinner';
+import LoadingSpinner from '../../components/LoadingSpinner';
 import Pagination from '../../components/Pagination';
 import Tabs from '../../govuk/Tabs';
 import TaskListMode from './TaskListMode';
